# 07 API Catalog/openapi/complaint-domain-service.v1.yaml
openapi: 3.0.3
info:
  title: ECMP Complaint Domain Service API
  version: 1.4.0
  description: |
    Complaint Domain Foundation + Processing + Assignment + Escalation + SLA REST API
    (CAPABILITY-004…008).
    Complaint is the business reason for a visit; QueueTicket is visit context only.
    One QueueTicket may link to many Complaints via queueTicketId.
    Independent bounded context — no Queue domain imports, no Workflow,
    Timeline, Comment, Attachment, Notification,
    Dashboard, Display, AI, or Authentication.
    Lifecycle: OPEN → IN_PROGRESS → RESOLVED → CLOSED; reopen RESOLVED → IN_PROGRESS.
    Assignment determines responsibility only — it does not change lifecycle status.
    At most one active Assignment per Complaint; history is append-only.
    Escalation determines handling level only — it does not change Assignment or status.
    At most one current Escalation per Complaint; history is append-only; level must increase.
    SLA calculates due/remaining/breach only — no scheduler, notification, or auto-escalation.
    At most one active ComplaintSLA per Complaint. Closing a Complaint completes active SLA.
  x-ear-id: API-390
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
  /api/v1/complaints:
    post:
      summary: Create complaint
      operationId: createComplaintDomain
      x-ear-id: API-390
      tags: [Complaint Domain]
      security: []
      description: |
        Create a Complaint linked to queueTicketId. Initial status OPEN.
        Does not invoke Queue operations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: List complaints by organization
      operationId: listComplaintDomain
      x-ear-id: API-391
      tags: [Complaint Domain]
      security: []
      parameters:
        - name: organizationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: Get complaint by id
      operationId: getComplaintDomain
      x-ear-id: API-392
      tags: [Complaint Domain]
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      summary: Update complaint
      operationId: updateComplaintDomain
      x-ear-id: API-393
      tags: [Complaint Domain]
      security: []
      description: |
        Update mutable fields. Status transitions must follow
        OPEN → IN_PROGRESS → RESOLVED → CLOSED (no OPEN → CLOSED).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete complaint
      operationId: deleteComplaintDomain
      x-ear-id: API-394
      tags: [Complaint Domain]
      security: []
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/start:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Start complaint processing
      operationId: startComplaintProcessing
      x-ear-id: API-397
      tags: [Complaint Domain]
      security: []
      description: |
        OPEN → IN_PROGRESS. Domain-validated lifecycle transition.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/resolve:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Resolve complaint
      operationId: resolveComplaint
      x-ear-id: API-398
      tags: [Complaint Domain]
      security: []
      description: |
        IN_PROGRESS → RESOLVED. Requires summary and resolvedBy.
        Creates a Resolution value object (immutable after CLOSED).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/close:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Close complaint
      operationId: closeComplaintDomain
      x-ear-id: API-399
      tags: [Complaint Domain]
      security: []
      description: |
        RESOLVED → CLOSED. Resolution becomes immutable.
        Request body is optional.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/reopen:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Reopen complaint
      operationId: reopenComplaint
      x-ear-id: API-400
      tags: [Complaint Domain]
      security: []
      description: |
        RESOLVED → IN_PROGRESS. Optional reason (not persisted — Timeline out of scope).
        Clears prior Resolution so a new resolve may be recorded.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReopenRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/assign:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Assign complaint
      operationId: assignComplaintDomain
      x-ear-id: API-401
      tags: [Complaint Domain]
      security: []
      description: |
        Create the first active Assignment for a Complaint.
        Rejects when an active assignment already exists (use reassign).
        Does not change Complaint lifecycle status.
        CAPABILITY-006 supports assigneeType USER only (TEAM/QUEUE/SYSTEM design-ready).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/reassign:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Reassign complaint
      operationId: reassignComplaintDomain
      x-ear-id: API-402
      tags: [Complaint Domain]
      security: []
      description: |
        Release the active Assignment and append a new active Assignment.
        History rows are not rewritten. Does not change Complaint status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/unassign:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Unassign complaint
      operationId: unassignComplaintDomain
      x-ear-id: API-403
      tags: [Complaint Domain]
      security: []
      description: |
        Release the active Assignment. Complaint has no active assignee.
        Does not change Complaint lifecycle status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnassignRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/assignment:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: Get current assignment
      operationId: getComplaintAssignmentDomain
      x-ear-id: API-404
      tags: [Complaint Domain]
      security: []
      description: Return the single active Assignment for a Complaint, if any.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/assignments:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: List assignment history
      operationId: listComplaintAssignmentsDomain
      x-ear-id: API-405
      tags: [Complaint Domain]
      security: []
      description: Append-only Assignment history for a Complaint (oldest first).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/escalate:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    post:
      summary: Escalate complaint
      operationId: escalateComplaintDomain
      x-ear-id: API-406
      tags: [Complaint Domain]
      security: []
      description: |
        Append a new current Escalation. Prior current becomes historical.
        Level must strictly increase when a current escalation exists.
        Does not change Assignment or Complaint lifecycle status.
        No auto-escalation / scheduler / notification in CAPABILITY-007.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/escalation:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: Get current escalation
      operationId: getComplaintEscalationDomain
      x-ear-id: API-407
      tags: [Complaint Domain]
      security: []
      description: Return the single current Escalation for a Complaint, if any.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/complaints/{complaintId}/escalations:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: List escalation history
      operationId: listComplaintEscalationsDomain
      x-ear-id: API-408
      tags: [Complaint Domain]
      security: []
      description: Append-only Escalation history for a Complaint (oldest first).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/complaints/{complaintId}/sla:
    parameters:
      - $ref: "#/components/parameters/ComplaintId"
    get:
      summary: Get complaint SLA
      operationId: getComplaintSlaDomain
      x-ear-id: API-412
      tags: [Complaint Domain]
      security: []
      description: |
        Return active ComplaintSLA when present, otherwise the latest
        completed/historical SLA. May refresh breach detection on active rows.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintSLADataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
...
    ComplaintDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ComplaintResponse"
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...